bin/requirements.txt
mysql.connector
numpy
pandas
requests
Faker